Implement `infer_schema!` with Macros 1.1
What's that RFC authors? You want to specifically limit the RFC to
stabilize only custom-derive and exclude bang macros? WE'LL SEE ABOUT
THAT!

I had to comment out one test from the syntex crate, as cargo won't let
me have a feature called dotenv while dotenv is a dev dependency. That
code will move to the shared crate soon, so it shouldn't be a problem.